[ui] Refactor file and folder structure and ensure consistent naming convention
The Shadcn UI components use "Kebab case" (words-separated-by-dash) convention. It makes for a more readable file names. This commit updates all other files to this same naming convention.
