+++ app/src/app/legal-units/[id]/@header/default.tsx
@@ -1,4 +1,4 @@
-import {getLegalUnitById} from "@/app/legal-units/[id]/requests";
+import {getLegalUnitById} from "@/app/legal-units/[id]/legal-unit-requests";
 
 export default async function LegalUnitHeader({params: {id}}: { readonly params: { id: string } }) {
   const unit = await getLegalUnitById(id);
