+++ app/src/app/legal-units/[id]/contact/page.tsx
@@ -2,7 +2,7 @@ import {Metadata} from "next";
 import {notFound} from "next/navigation";
 import {Separator} from "@/components/ui/separator";
 import ContactInfoForm from "@/app/legal-units/[id]/contact/contact-info-form";
-import {getLegalUnitById} from "@/app/legal-units/[id]/requests";
+import {getLegalUnitById} from "@/app/legal-units/[id]/legal-unit-requests";
 
 export const metadata: Metadata = {
   title: "Legal Unit | Contact"
