+++ app/src/app/search/components/reset-filter-button.tsx
