+++ app/src/app/search/components/search-result-table.tsx
