+++ app/src/components/error-box.tsx
