+++ app/src/components/info-box.tsx
