+++ app/src/components/profile-avatar.tsx
